Almost Done
Animations added, mouse hovering and changing to active mouse cursor works. Also the animations are rushed and look kind of bad. Apart from that, the translation of "woof woof" to actual human words is under way, with the setting page built out and changing a universal variable from true to false and false to true; just need to add the if statements to dialogue code so it syays woof when translate is false, also there's still no audio.